uint_t implemented

+++ main.c
@@ -1,12 +1,13 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include "binaryset.h"
 
 int main() {
     BinaryNode_t *binaryHeadNode = NULL;
     BinaryNode_t *binaryTemporaryNode;
-    int setLen, isUnique, decimalNum;
-    int i = 0;
+    uint8_t setLen, isUnique, decimalNum;
+    uint8_t i = 0;
     char setLenStr[DATA_SIZE],data[DATA_SIZE], binaryNum[DATA_SIZE];
     do {
         puts("Enter size of binary digit set");
